Fix: Create if statement so race with id of 1 is skipped
There is a current bug in the backend that causes an error when the http://localhost:8000/api/races/1/start endpoint is tried to be reached. Until Udacity can fix this issue,
I have created an if statement that skips over the race id of 1 to avoid this endpoint. A new race is created with the same track and player id but a new race id before the
startRace API is called.